Move vec-math utilities to shared src/engine/math/
