+++ src/engine/math/vec-math.ts
