+++ src/engine/physics/execution.ts
@@ -11,7 +11,7 @@
 import type { Vec3 } from "../types/index.js";
 import type { PhysicsConfig } from "./constants.js";
 import type { Rng } from "../rng.js";
-import { clamp } from "./vec-math.js";
+import { clamp } from "../math/vec-math.js";
 
 /**
  * Compute execution quality (0-1) for a rally shot.
