+++ src/engine/player/serve-decision.ts
@@ -8,7 +8,7 @@
 import type { ServeContext, ServeIntention } from "../types/index.js";
 import type { Rng } from "../rng.js";
 import type { PlayerEngineConfig } from "./constants.js";
-import { clamp } from "../physics/vec-math.js";
+import { clamp } from "../math/vec-math.js";
 import {
   mapTendency,
   mapTendencyWithRisk,
